folder-and-files-list
The two functions give you the names of the folders and files.

current-folder-and-files-names-list.js creates a txt file that has a list of of the names of the folders and files in the designated folder path.

folder-and-files-names-list creates a txt file that has a list of the names of the folders and files in all subfolders